moves shared util functions to lib
