+++ install/install_chrome.py
@@ -1,6 +1,6 @@
 from exceptions import InstallerException
 from install.ChromeInstaller import ChromeInstaller
-from util import create_logger
+from lib.util import create_logger
 
 
 def install_chrome(debug=False):
@@ -10,6 +10,7 @@ def install_chrome(debug=False):
         logger.info('Getting Chrome version: {}'.format(installer.version))
         logger.info('Downloading and installing browser...')
         installer.download_and_install_browser()
+        logger.info('On OSX, drag the browser icon into the Application folder using the popup.')
         logger.info('Downloading and unpacking driver...')
         installer.download_driver()
     except InstallerException as e:
