+++ lib/util.py
