+++ merge/SheetManager.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 from os import path
 
 from exceptions import SheetManagerException
-from util import create_new_filename
+from lib.util import create_new_filename
 
 SEP = '-' * 70
